backend_noticias/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1
# praw==7.7.1  # Opcional para API de Reddit
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.25.0